+++ components/aboutMe/AboutMe.constants.js
@@ -1,0 +1,9 @@
+const ARRAY_OF_LINK = [
+    'https://s1.iconbird.com/ico/2013/9/446/w512h5121380376407MetroUIHTML5.png',
+    'https://cdn.icon-icons.com/icons2/2107/PNG/512/file_type_css_icon_130661.png',
+    'https://img2.freepng.ru/20180411/cvq/kisspng-javascript-html-computer-software-web-browser-watermark-5acdbd54ac19f7.4484983215234327887049.jpg',
+    'https://reactnative.dev/img/tiny_logo.png'
+
+];
+
+export default ARRAY_OF_LINK;
