+++ res/colors.js
@@ -1,0 +1,7 @@
+const colors = {
+    white: 'white',
+    purple: '#5e29e3',
+    black: "black"
+}
+
+export default colors;
